Move `pkg`-specific code into a new `pkg` submodule
